move code out of mod.rs into sub modules
