+++ storage/src/mmr/benches/bench.rs
@@ -1,11 +1,11 @@
 use criterion::criterion_main;
 
-mod build_mmr;
+mod build;
 mod prove_element_range;
 mod prove_single_element;
 
 criterion_main!(
-    build_mmr::benches,
+    build::benches,
     prove_element_range::benches,
     prove_single_element::benches,
 );
