+++ storage/src/mmr/iterator.rs
@@ -1,0 +1,137 @@
+/// A PeakIterator returns a (position, height) tuple for each peak in an MMR with the given size,
+/// in decreasing order of height.
+///
+/// For the example MMR depicted at the top of this file, the PeakIterator would yield:
+/// ```text
+/// [(14, 3), (17, 1), (18, 0)]
+/// ```
+#[derive(Default)]
+pub(crate) struct PeakIterator {
+    size: u64,     // number of nodes in the MMR at the point the iterator was initialized
+    node_pos: u64, // position of the current node
+    two_h: u64,    // 2^(height+1) of the current node
+}
+
+impl PeakIterator {
+    /// Return a new PeakIterator over the peaks of a MMR with the given number of nodes.
+    pub(crate) fn new(size: u64) -> PeakIterator {
+        if size == 0 {
+            return PeakIterator::default();
+        }
+        // Compute the position at which to start the search for peaks. This starting position will
+        // not be in the MMR unless it happens to be a single perfect binary tree, but that's OK as
+        // we will descend leftward until we find the first peak.
+        let start = u64::MAX >> size.leading_zeros();
+        let two_h = 1 << start.trailing_ones();
+        PeakIterator {
+            size,
+            node_pos: start - 1,
+            two_h,
+        }
+    }
+}
+
+impl Iterator for PeakIterator {
+    type Item = (u64, u32); // (peak, height)
+
+    fn next(&mut self) -> Option<Self::Item> {
+        while self.two_h > 1 {
+            if self.node_pos < self.size {
+                // found a peak
+                let peak_item = (self.node_pos, self.two_h.trailing_zeros() - 1);
+                // move to the right sibling
+                self.node_pos += self.two_h - 1;
+                assert!(self.node_pos >= self.size); // sibling shouldn't be in the MMR if MMR is valid
+                return Some(peak_item);
+            }
+            // descend to the left child
+            self.two_h >>= 1;
+            self.node_pos -= self.two_h;
+        }
+        None
+    }
+}
+
+/// Returns the set of peaks that will require a new parent after adding the next leaf to an MMR
+/// with the given peaks. This set is non-empty only if there is a height-0 (leaf) peak in the MMR.
+/// The result will contain this leaf peak plus the other MMR peaks with contiguously increasing
+/// height. Nodes in the result are ordered by decreasing height.
+pub(crate) fn nodes_needing_parents(peak_iterator: PeakIterator) -> Vec<u64> {
+    let mut peaks = Vec::new();
+    let mut last_height = u32::MAX;
+
+    for (peak_pos, height) in peak_iterator {
+        assert!(last_height > 0);
+        assert!(height < last_height);
+        if height != last_height - 1 {
+            peaks.clear();
+        }
+        peaks.push(peak_pos);
+        last_height = height;
+    }
+    if last_height != 0 {
+        // there is no peak that is a leaf
+        peaks.clear();
+    }
+    peaks
+}
+
+/// A PathIterator returns a (parent_pos, sibling_pos) tuple for the sibling of each node along the
+/// path from a given perfect binary tree peak to a designated leaf, not including the peak itself.
+///
+/// For example, consider the tree below and the path from the peak to leaf node 3. Nodes on the
+/// path are [6, 5, 3] and tagged with '*' in the diagram):
+///
+/// ```text
+///
+///          6*
+///        /   \
+///       2     5*
+///      / \   / \
+///     0   1 3*  4
+///
+/// A PathIterator for this example yields:
+///    [(6, 2), (5, 4)]
+/// ```
+#[derive(Debug)]
+pub(crate) struct PathIterator {
+    leaf_pos: u64, // position of the leaf node in the path
+    node_pos: u64, // current node position in the path from peak to leaf
+    two_h: u64,    // 2^height of the current node
+}
+
+impl PathIterator {
+    /// Return a PathIterator over the siblings of nodes along the path from peak to leaf in the
+    /// perfect binary tree with peak `peak_pos` and having height `height`, not including the peak
+    /// itself.
+    pub(crate) fn new(leaf_pos: u64, peak_pos: u64, height: u32) -> PathIterator {
+        PathIterator {
+            leaf_pos,
+            node_pos: peak_pos,
+            two_h: 1 << height,
+        }
+    }
+}
+
+impl Iterator for PathIterator {
+    type Item = (u64, u64); // (parent_pos, sibling_pos)
+
+    fn next(&mut self) -> Option<Self::Item> {
+        if self.two_h <= 1 {
+            return None;
+        }
+
+        let left_pos = self.node_pos - self.two_h;
+        let right_pos = left_pos + self.two_h - 1;
+        self.two_h >>= 1;
+
+        if left_pos < self.leaf_pos {
+            let r = Some((self.node_pos, left_pos));
+            self.node_pos = right_pos;
+            return r;
+        }
+        let r = Some((self.node_pos, right_pos));
+        self.node_pos = left_pos;
+        r
+    }
+}
