+++ storage/src/mmr/mem.rs
@@ -1,7 +1,9 @@
 //! A bare-bones MMR structure without pruning and where all nodes are hashes & maintained in
 //! memory within a single vector.
 
-use crate::mmr::{nodes_needing_parents, Hasher, PathIterator, PeakIterator, Proof};
+use crate::mmr::hasher::Hasher;
+use crate::mmr::iterator::{nodes_needing_parents, PathIterator, PeakIterator};
+use crate::mmr::verification::Proof;
 use commonware_cryptography::{Digest, Hasher as CHasher};
 
 /// Implementation of `Mmr`.
@@ -151,9 +153,10 @@ impl<H: CHasher> Mmr<H> {
 
 #[cfg(test)]
 mod tests {
+    use crate::mmr::hasher::Hasher;
+    use crate::mmr::iterator::nodes_needing_parents;
     use crate::mmr::mem::Mmr;
-    use crate::mmr::{nodes_needing_parents, verify_proof, verify_range_proof, Hasher};
-    use commonware_cryptography::{Digest, Hasher as CHasher, Sha256};
+    use commonware_cryptography::{Digest, Sha256};
 
     #[test]
     /// Test MMR building by consecutively adding 11 equal elements to a new MMR, producing the
@@ -236,259 +239,5 @@ mod tests {
         let peak_hashes = [hash14, hash17, mmr.nodes[18].clone()];
         let expected_root_hash = mmr_hasher.root_hash(19, peak_hashes.iter());
         assert_eq!(root_hash, expected_root_hash, "incorrect root hash");
-
-        // confirm the proof of inclusion for each leaf successfully verifies
-        for leaf in leaves.iter().by_ref() {
-            let proof = mmr.proof(*leaf);
-            assert!(
-                verify_proof::<Sha256>(&proof, &element, *leaf, &root_hash, &mut hasher),
-                "valid proof should verify successfully"
-            );
-        }
-
-        // confirm mangling the proof or proof args results in failed validation
-        const POS: u64 = 18;
-        let proof = mmr.proof(POS);
-        assert!(
-            verify_proof::<Sha256>(&proof, &element, POS, &root_hash, &mut hasher),
-            "proof verification should be successful"
-        );
-        assert!(
-            !verify_proof::<Sha256>(&proof, &element, POS + 1, &root_hash, &mut hasher),
-            "proof verification should fail with incorrect element position"
-        );
-        assert!(
-            !verify_proof::<Sha256>(&proof, &element, POS - 1, &root_hash, &mut hasher),
-            "proof verification should fail with incorrect element position 2"
-        );
-        assert!(
-            !verify_proof::<Sha256>(
-                &proof,
-                &Digest::from(vec![0u8; Sha256::len()]),
-                POS,
-                &root_hash,
-                &mut hasher
-            ),
-            "proof verification should fail with mangled element"
-        );
-        let root_hash2 = Digest::from(vec![0u8; Sha256::len()]);
-        assert!(
-            !verify_proof::<Sha256>(&proof, &element, POS, &root_hash2, &mut hasher),
-            "proof verification should fail with mangled root_hash"
-        );
-        let mut proof2 = proof.clone();
-        proof2.hashes[0] = Digest::from(vec![0u8; Sha256::len()]);
-        assert!(
-            !verify_proof::<Sha256>(&proof2, &element, POS, &root_hash, &mut hasher),
-            "proof verification should fail with mangled proof hash"
-        );
-        proof2 = proof.clone();
-        proof2.size = 10;
-        assert!(
-            !verify_proof::<Sha256>(&proof2, &element, POS, &root_hash, &mut hasher),
-            "proof verification should fail with incorrect size"
-        );
-        proof2 = proof.clone();
-        proof2.hashes.push(Digest::from(vec![0u8; Sha256::len()]));
-        assert!(
-            !verify_proof::<Sha256>(&proof2, &element, POS, &root_hash, &mut hasher),
-            "proof verification should fail with extra hash"
-        );
-        proof2 = proof.clone();
-        while !proof2.hashes.is_empty() {
-            proof2.hashes.pop();
-            assert!(
-                !verify_proof::<Sha256>(&proof2, &element, 7, &root_hash, &mut hasher),
-                "proof verification should fail with missing hashes"
-            );
-        }
-        proof2 = proof.clone();
-        proof2.hashes.clear();
-        proof2
-            .hashes
-            .extend(proof.hashes[0..peak_hashes.len() - 1].iter().cloned());
-        // sneak in an extra hash that won't be used in the computation and make sure it's detected
-        proof2.hashes.push(Digest::from(vec![0u8; Sha256::len()]));
-        proof2
-            .hashes
-            .extend(proof.hashes[peak_hashes.len() - 1..].iter().cloned());
-        assert!(
-            !verify_proof::<Sha256>(&proof2, &element, POS, &root_hash, &mut hasher),
-            "proof verification should fail with extra hash even if it's unused by the computation"
-        );
-    }
-
-    #[test]
-    fn test_range_proofs() {
-        // create a new MMR and add a non-trivial amount (47) of elements
-        let mut mmr: Mmr<Sha256> = Mmr::default();
-        let mut elements = Vec::<Digest>::new();
-        let mut element_positions = Vec::<u64>::new();
-        for i in 0..49 {
-            elements.push(Digest::from(vec![i as u8; Sha256::len()]));
-            element_positions.push(mmr.add(elements.last().unwrap()));
-        }
-        // test range proofs over all possible ranges of at least 2 elements
-        let root_hash = mmr.root_hash();
-        let mut hasher = Sha256::default();
-        for i in 0..elements.len() {
-            for j in i + 1..elements.len() {
-                let start_pos = element_positions[i];
-                let end_pos = element_positions[j];
-                let range_proof = mmr.range_proof(start_pos, end_pos);
-                assert!(
-                    verify_range_proof::<Sha256>(
-                        &range_proof,
-                        &elements[i..j + 1],
-                        start_pos,
-                        end_pos,
-                        &root_hash,
-                        &mut hasher,
-                    ),
-                    "valid range proof should verify successfully {}:{}",
-                    i,
-                    j
-                );
-            }
-        }
-
-        // create a test range for which we will mangle data and confirm the proof fails
-        let start_index = 33;
-        let end_index = 39;
-        let start_pos = element_positions[start_index];
-        let end_pos = element_positions[end_index];
-        let range_proof = mmr.range_proof(start_pos, end_pos);
-        let valid_elements = &elements[start_index..end_index + 1];
-        assert!(
-            verify_range_proof::<Sha256>(
-                &range_proof,
-                valid_elements,
-                start_pos,
-                end_pos,
-                &root_hash,
-                &mut hasher,
-            ),
-            "valid range proof should verify successfully"
-        );
-        let mut invalid_proof = range_proof.clone();
-        for _i in 0..range_proof.hashes.len() {
-            invalid_proof.hashes.remove(0);
-            assert!(
-                !verify_range_proof::<Sha256>(
-                    &range_proof,
-                    &Vec::new(),
-                    start_pos,
-                    end_pos,
-                    &root_hash,
-                    &mut hasher,
-                ),
-                "range proof with removed elements should fail"
-            );
-        }
-        // confirm proof fails with invalid element hashes
-        for i in 0..elements.len() {
-            for j in i..elements.len() {
-                assert!(
-                    (i == start_index && j == end_index) // exclude the valid element range
-                        || !verify_range_proof::<Sha256>(
-                            &range_proof,
-                            &elements[i..j + 1],
-                            start_pos,
-                            end_pos,
-                            &root_hash,
-                            &mut hasher,
-                        ),
-                    "range proof with invalid elements should fail {}:{}",
-                    i,
-                    j
-                );
-            }
-        }
-        // confirm proof fails with invalid root hash
-        let mut invalid_root_hash = vec![0; Sha256::len()];
-        invalid_root_hash[29] = root_hash[29] + 1;
-        assert!(
-            !verify_range_proof::<Sha256>(
-                &range_proof,
-                valid_elements,
-                start_pos,
-                end_pos,
-                &Digest::from(invalid_root_hash),
-                &mut hasher,
-            ),
-            "range proof with invalid proof should fail"
-        );
-        // mangle the proof and confirm it fails
-        let mut invalid_proof = range_proof.clone();
-        invalid_proof.hashes[1] = Digest::from(vec![0u8; Sha256::len()]);
-        assert!(
-            !verify_range_proof::<Sha256>(
-                &invalid_proof,
-                valid_elements,
-                start_pos,
-                end_pos,
-                &root_hash,
-                &mut hasher,
-            ),
-            "mangled range proof should fail verification"
-        );
-        // inserting elements into the proof should also cause it to fail (malleability check)
-        for i in 0..range_proof.hashes.len() {
-            let mut invalid_proof = range_proof.clone();
-            invalid_proof
-                .hashes
-                .insert(i, Digest::from(vec![0u8; Sha256::len()]));
-            assert!(
-                !verify_range_proof::<Sha256>(
-                    &invalid_proof,
-                    valid_elements,
-                    start_pos,
-                    end_pos,
-                    &root_hash,
-                    &mut hasher,
-                ),
-                "mangled range proof should fail verification. inserted element at: {}",
-                i
-            );
-        }
-        // removing proof elements should cause verification to fail
-        let mut invalid_proof = range_proof.clone();
-        for _ in 0..range_proof.hashes.len() {
-            invalid_proof.hashes.remove(0);
-            assert!(
-                !verify_range_proof::<Sha256>(
-                    &invalid_proof,
-                    valid_elements,
-                    start_pos,
-                    end_pos,
-                    &root_hash,
-                    &mut hasher,
-                ),
-                "shortened range proof should fail verification"
-            );
-        }
-        // bad element range should cause verification to fail
-        for i in 0..elements.len() {
-            for j in 0..elements.len() {
-                let start_pos2 = element_positions[i];
-                let end_pos2 = element_positions[j];
-                if start_pos2 == start_pos && end_pos2 == end_pos {
-                    continue;
-                }
-                assert!(
-                    !verify_range_proof::<Sha256>(
-                        &range_proof,
-                        valid_elements,
-                        start_pos2,
-                        end_pos2,
-                        &root_hash,
-                        &mut hasher,
-                    ),
-                    "bad element range should fail verification {}:{}",
-                    i,
-                    j
-                );
-            }
-        }
     }
 }
